refactor: rely on same queryParams
